+++ src/FamilyHubs.ServiceDirectory.Web/Models/Interfaces/IPostFilter.cs
@@ -1,0 +1,6 @@
+﻿namespace FamilyHubs.ServiceDirectory.Web.Models.Interfaces;
+
+public interface IPostFilter : IFilter
+{
+    public string? Value { get; }
+}
