+++ src/FamilyHubs.ServiceDirectory.Web/Pages/_ViewImports.cshtml
@@ -1,4 +1,5 @@
 ﻿@using FamilyHubs.ServiceDirectory.Web
 @using FamilyHubs.ServiceDirectory.Web.Models
+@using FamilyHubs.ServiceDirectory.Web.Models.Interfaces
 @namespace FamilyHubs.ServiceDirectory.Web.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
